+++ src/lib/Prefs.js
@@ -324,6 +324,18 @@ var Prefs = class
         );
 
         this._setKey(
+            'visibility',
+            'weather',
+            'GtkSwitch',
+            true,
+            {
+                default: true,
+                minimal: true,
+                superminimal: true,
+            }
+        );
+
+        this._setKey(
             'icons',
             'app-menu-icon',
             'GtkSwitch',